partial project restructure
